app/backend/requirements.txt
dill==0.3.9
Flask==3.0.3
Flask_Cors==4.0.0
huggingface_hub==0.25.1
langchain==0.3.3
langchain_community==0.3.2
numpy==2.1.2
pandas==2.2.3
peft==0.13.0
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.5.2
tensorflow==2.17.0
tensorflow_intel==2.17.0
torch==2.3.0
transformers==4.45.1
